requirements.txt
# OpenMontage - Core Dependencies
pyyaml>=6.0
pydantic>=2.0
jsonschema>=4.20
python-dotenv>=1.0
